requirements.txt
streamlit
pandas
matplotlib
pillow
opencv-python
pytesseract
